pyproject.toml
[project]
name = "my-db-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.3",
    "mcp[cli]>=1.3.0",
    "mysqlclient>=2.2.7",
    "openai>=1.82.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.0",
]
